Separate test and src files
