+++ server/src/game_types.c
