+++ server/src/game_types.h
