+++ server/src/util.c
