+++ server/test/game_logic_test.h
@@ -1,0 +1,6 @@
+#ifndef GAME_LOGIC_TEST_H
+#define GAME_LOGIC_TEST_H
+
+void test_evaluate_word_challenge_guess(void);
+
+#endif // !GAME_LOGIC_TEST_H
