+++ server/test/hash_table_test.h
@@ -1,0 +1,7 @@
+#ifndef HASH_TABLE_TEST_H
+#define HASH_TABLE_TEST_H
+
+void test_hash_table(void);
+void test_call_HT_delete_on_empty_hash_table(void);
+
+#endif // !HASH_TABLE_TEST_H
