+++ server/test/test_util.c
@@ -1,3 +1,4 @@
+#include "test_util.h"
 #include <stdio.h>
 #include <unistd.h>
