+++ server/test/timer_test.h
@@ -1,0 +1,9 @@
+#ifndef TIMER_TEST_H
+#define TIMER_TEST_H
+
+void test_timer_lifecycle(void);
+void test_timer_list_examine(void);
+void test_timer_rearm(void);
+void test_timer_arm_within_examine(void);
+
+#endif // !TIMER_TEST_H
