+++ server/test/util_test.c
@@ -1,9 +1,9 @@
-#include "util.h"
+#include "util_test.h"
+#include "../src/util.h"
+
 #include <assert.h>
 #include <stdlib.h>
 #include <string.h>
-
-static void test_generate_random_string(void);
 
 void test_generate_random_string(void) {
   char *s = generate_random_string(7);
