+++ server/test/util_test.h
@@ -1,0 +1,6 @@
+#ifndef UTIL_TEST_H
+#define UTIL_TEST_H
+
+void test_generate_random_string(void);
+
+#endif // !UTIL_TEST_H
